requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
cssbeautifier==1.14.7
cymem==2.0.7
Django==4.1.7
django-cacheops==6.2
django-sass-processor==1.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djlint==1.19.16
dnspython==2.2.0
EditorConfig==0.12.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.5.0/es_core_news_sm-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
funcy==1.18
googlemaps==4.6.0
greenlet==1.1.2
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsbeautifier==1.14.7
langcodes==3.3.0
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
openpyxl==3.0.9
packaging==23.0
pandas==1.4.1
pathspec==0.11.0
pathy==0.10.1
peewee==3.14.9
pi==0.1.2
preshed==3.0.8
pycodestyle==2.8.0
pydantic==1.10.5
PyJWT==2.6.0
pymongo==4.0.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2021.3
PyYAML==6.0
redis==4.5.1
regex==2022.10.31
requests==2.28.2
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.31
sqlparse==0.4.3
srsly==2.4.6
thinc==8.1.7
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
wasabi==1.1.1
Werkzeug==2.2.3
WTForms==3.0.1
yarl==1.8.2
zipp==3.15.0
